test: use a real fixture instead of an inline file for the malformed source map test
