+++ test/e2e/fixtures/malformed-source-map/malformed-source-map.js
@@ -1,0 +1,5 @@
+//# sourceMappingURL=data:application/json;base64,bm90LWpzb24=
+
+export default function testMalformedSourceMap() {
+  throw new Error('original module error')
+}
