+++ test/e2e/fixtures/malformed-source-map/some-test.spec.ts
@@ -1,0 +1,10 @@
+import { expect, test } from 'vitest'
+import testMalformedSourceMap from './malformed-source-map.js'
+
+test('passing test remains visible', () => {
+  expect(true).toBe(true)
+})
+
+test('reports the original module error', () => {
+  testMalformedSourceMap()
+})
